+++ packages/maker/deb/package.json
@@ -8,11 +8,13 @@
   "main": "dist/MakerDeb.js",
   "typings": "dist/MakerDeb.d.ts",
   "devDependencies": {
+    "@electron-forge/test-utils": "7.6.0",
     "chai": "^4.3.3",
     "chai-as-promised": "^7.0.0",
     "mocha": "^9.0.1",
     "proxyquire": "^2.1.3",
-    "sinon": "^13.0.1"
+    "sinon": "^13.0.1",
+    "vitest": "^2.1.6"
   },
   "engines": {
     "node": ">= 16.4.0"
